+++ test/test_HP3ParClient_host.py
@@ -14,7 +14,7 @@
 
 """Test class of 3Par Client handling Host."""
 
-import HP3ParClient_base as hp3parbase
+from test import HP3ParClient_base as hp3parbase
 
 from hp3parclient import exceptions