+++ test/test_HP3ParClient_retry.py
@@ -18,7 +18,7 @@ import importlib
 import mock
 import requests
 
-import HP3ParClient_base as hp3parbase
+from test import HP3ParClient_base as hp3parbase
 
 from hp3parclient import exceptions