+++ test/test_HP3ParClient_volume.py
@@ -16,7 +16,7 @@
 
 from testconfig import config
 
-import HP3ParClient_base as hp3parbase
+from test import HP3ParClient_base as hp3parbase
 
 from hp3parclient import exceptions
 
@@ -397,6 +397,43 @@ class HP3ParClientVolumeTestCase(hp3parbase.HP3ParClientBaseTestCase):
         )
 
         self.printFooter('copy_volume')
+
+    def test_7_copy_volume_failure(self):
+        self.printHeader('copy_volume_failure')
+
+        # add one
+        optional = {'comment': 'test volume', 'tpvv': True,
+                    'snapCPG': CPG_NAME1}
+        self.cl.createVolume(VOLUME_NAME1, CPG_NAME1, SIZE, optional)
+        self.cl.createVolume(VOLUME_NAME2, CPG_NAME1, SIZE, optional)
+
+        optional = {'online': False, 'tpvv': True}
+        self.assertRaises(
+            exceptions.HTTPBadRequest,
+            self.cl.copyVolume,
+            VOLUME_NAME1,
+            VOLUME_NAME2,
+            CPG_NAME1,
+            optional)
+
+        optional = {'online': False, 'tpdd': True}
+        self.assertRaises(
+            exceptions.HTTPBadRequest,
+            self.cl.copyVolume,
+            VOLUME_NAME1,
+            VOLUME_NAME2,
+            CPG_NAME1,
+            optional)
+
+        # destCPG isn't allowed to go to the 3PAR during an
+        # offline copy.  The client strips it out, so this should pass
+        optional = {'online': False, 'destCPG': 'test'}
+        self.cl.copyVolume(VOLUME_NAME1, VOLUME_NAME2, CPG_NAME1, optional)
+        self.cl.getVolume(VOLUME_NAME2)
+        self.cl.deleteVolume(VOLUME_NAME2)
+        self.cl.deleteVolume(VOLUME_NAME1)
+
+        self.printFooter('copy_volume_failure')
 
     def test_7_create_volume_set(self):
         self.printHeader('create_volume_set')